Nutri Project Sem 6/requirement.txt
flask
flask-cors
pandas
scipy
python-dotenv
langchain-groq
groq
gtts
fpdf
plotly